+++ glean-core/src/database/sqlite/schema.rs
@@ -16,7 +16,7 @@ use super::connection::ConnectionOpener;
 pub struct Schema;
 
 impl ConnectionOpener for Schema {
-    const MAX_SCHEMA_VERSION: u32 = 2;
+    const MAX_SCHEMA_VERSION: u32 = 3;
 
     type Error = SchemaError;
 
@@ -53,15 +53,24 @@ impl ConnectionOpener for Schema {
     fn create(tx: &mut Transaction<'_>) -> Result<(), Self::Error> {
         tx.execute_batch(
             "
-             CREATE TABLE telemetry(
-               id TEXT NOT NULL,
-               ping TEXT NOT NULL,
-               lifetime TEXT NOT NULL,
-               labels TEXT NOT NULL, -- can't be null or ON CONFLICT won't work
-               value BLOB,
-               UNIQUE(id, ping, labels)
-             );
-             CREATE TABLE migration(id INTEGER PRIMARY KEY, state TEXT NOT NULL);
+            CREATE TABLE telemetry(
+                id TEXT NOT NULL,
+                ping TEXT NOT NULL,
+                lifetime TEXT NOT NULL,
+                labels TEXT NOT NULL, -- can't be null or ON CONFLICT won't work
+                value BLOB,
+                UNIQUE(id, ping, labels)
+            );
+            CREATE TABLE migration(id INTEGER PRIMARY KEY, state TEXT NOT NULL);
+            CREATE TABLE submitted_pings(
+                document_id TEXT PRIMARY KEY,
+                ping TEXT NOT NULL,
+                date_submitted INTEGER NOT NULL,
+                date_uploaded INTEGER,
+                upload_failed BOOLEAN NOT NULL,
+                payload TEXT
+            );
+            CREATE INDEX submitted_pings_ping on submitted_pings(ping);
             ",
         )?;
         Ok(())
@@ -88,6 +97,24 @@ impl ConnectionOpener for Schema {
                     log::info!("Client ID already exists. Marking migration as done.");
                     tx.execute("INSERT INTO migration (id, state) VALUES (1, 'done') ON CONFLICT(id) DO UPDATE SET state = excluded.state", [])?;
                 }
+                Ok(())
+            }
+            3 => {
+                log::info!("Upgrading user_version to 3");
+                // Clients upgrading to schema 3 don't have the table or index
+                tx.execute_batch(
+                    "
+                    CREATE TABLE submitted_pings(
+                        document_id TEXT PRIMARY KEY,
+                        ping TEXT NOT NULL,
+                        date_submitted INTEGER NOT NULL,
+                        date_uploaded INTEGER,
+                        upload_failed BOOLEAN NOT NULL,
+                        payload TEXT
+                    );
+                    CREATE INDEX submitted_pings_ping on submitted_pings(ping);
+                    ",
+                )?;
                 Ok(())
             }
             to_version => Err(SchemaError::UnsupportedSchemaVersion(to_version)),