+++ glean-core/src/lib_unit_tests.rs
@@ -238,6 +238,7 @@ fn experimentation_id_is_set_correctly() {
         session_sample_rate: 1.0,
         session_inactivity_timeout_ms: 1_800_000,
         events_ping_acceleration_factor: None,
+        enable_store_submitted_pings: false,
     })
     .unwrap();