+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased changes
 
+* Add mechanisms for storing and retrieving submitted pings ([#3585](https://github.com/mozilla/glean/pull/3585)).
+    * Add new `submitted_pings` table to the SQLite database.
+    * Add methods to store, retrieve, update, and clear stored submitted pings.
+    * Update Ping and uploader implementations to store and update submitted pings as appropriate.
+
 [Full changelog](https://github.com/mozilla/glean/compare/v70.0.0...main)
 
 # v70.0.0 (2026-08-20)
@@ -22,8 +27,8 @@
     * Implement glean-noop as a feature of glean-sym ([#3541](https://github.com/mozilla/glean/pull/3541))
     * Support pings ([#3544](https://github.com/mozilla/glean/pull/3544))
     * Implement the event metric ([#3534](https://github.com/mozilla/glean/pull/3534))
-    * BREAKING CHANGE: Switch from a noop feature to an `active` feature ([#3583](https://github.com/mozilla/glean/pull/3583))
 * iOS
+    * BREAKING CHANGE: Switch from a noop feature to an `active` feature ([#3583](https://github.com/mozilla/glean/pull/3583))
   * Implement the custom distribution metric type ([#3572](https://github.com/mozilla/glean/pull/3572))
 * Python
   * Implement the custom distribution metric type ([#3572](https://github.com/mozilla/glean/pull/3572))